requirements.txt
vanna==0.0.30
PyYAML==6.0
numpy==1.26.1
pandas==2.1.3
slack-sdk==3.26.1
